refactor: move search filters components to categories module
