+++ modules/categories/components/BreadcrumbNavigation.tsx
@@ -6,7 +6,7 @@ import {
   BreadcrumbList,
   BreadcrumbPage,
   BreadcrumbSeparator,
-} from "../ui/breadcrumb";
+} from "@/components/ui/breadcrumb";
 import Link from "next/link";
 
 interface Props {
