+++ modules/categories/components/CategoriesSidebar.tsx
