+++ modules/categories/components/SearchFilters.tsx
