+++ modules/categories/components/SearchInput.tsx
@@ -1,10 +1,10 @@
 "use client";
 
 import { BookmarkCheckIcon, ListFilterIcon, SearchIcon } from "lucide-react";
-import { Input } from "../ui/input";
+import { Input } from "@/components/ui/input";
 import CategoriesSidebar from "./CategoriesSidebar";
 import { useState } from "react";
-import { Button } from "../ui/button";
+import { Button } from "@/components/ui/button";
 import { useTRPC } from "@/trpc/client";
 import { useQuery } from "@tanstack/react-query";
 import Link from "next/link";
